// src/models/ReportTypeEnum.ts
export enum ReportType {
    SPAMMING_ADVERTISING_BEGGING = "SPAMMING_ADVERTISING_BEGGING",
    POSTING_NSFW_CONTENT_AS_SFW = "POSTING_NSFW_CONTENT_AS_SFW",
    IMPERSONATING_A_MODERATOR = "IMPERSONATING_A_MODERATOR",
    PIRATED_CONTENT_TORRENTS = "PIRATED_CONTENT_TORRENTS",
    POSTING_ONION_LINK = "POSTING_ONION_LINK",
    CRIMINAL_INSTRUCTIONS = "CRIMINAL_INSTRUCTIONS",
    DOXXING_HARASSMENT_STALKING = "DOXXING_HARASSMENT_STALKING",
    NON_CONSENSUAL_SEXUAL_IMAGES = "NON_CONSENSUAL_SEXUAL_IMAGES",
    INCITEMENT_TO_LAWLESS_ACTION = "INCITEMENT_TO_LAWLESS_ACTION",
    TRUE_THREAT = "TRUE_THREAT",
    ATTEMPTING_TO_TRIGGER_EPILEPSY = "ATTEMPTING_TO_TRIGGER_EPILEPSY",
    POSTING_IN_A_ROOM_DEDICATED_TO_CRIMINAL_ACTIVITY = "POSTING_IN_A_ROOM_DEDICATED_TO_CRIMINAL_ACTIVITY",
    UNDERAGE = "UNDERAGE",
    ILLEGAL_CONTENT = "ILLEGAL_CONTENT",
    REAL_CHILD_SEXUAL_ABUSE_MATERIAL = "REAL_CHILD_SEXUAL_ABUSE_MATERIAL",
    FICTIONAL_CHILD_SEXUAL_ABUSE_MATERIAL = "FICTIONAL_CHILD_SEXUAL_ABUSE_MATERIAL",
    OTHER_REAL_OBSCENE_PORNOGRAPHY = "OTHER_REAL_OBSCENE_PORNOGRAPHY",
    OTHER_FICTIONAL_OBSCENE_PORNOGRAPHY = "OTHER_FICTIONAL_OBSCENE_PORNOGRAPHY",
    SNUFF_CRUSH_MATERIAL = "SNUFF_CRUSH_MATERIAL",
    INAPPROPRIATE_COMMUNICATION_WITH_MINOR = "INAPPROPRIATE_COMMUNICATION_WITH_MINOR",
    KNOWINGLY_SOLICITING_IDENTIFYING_INFO_OR_SEXUAL_CONTENT_FROM_MINOR = "KNOWINGLY_SOLICITING_IDENTIFYING_INFO_OR_SEXUAL_CONTENT_FROM_MINOR",
    KNOWINGLY_SENDING_SEXUAL_MESSAGES_OR_CONTENT_TO_MINOR = "KNOWINGLY_SENDING_SEXUAL_MESSAGES_OR_CONTENT_TO_MINOR"
}